# docs/swagger.yaml
definitions:
...
  task.DispatchTaskRequest:
    properties:
      end_time:
        type: string
      start_time:
        type: string
      task_type:
        type: string
      user_id:
        example: "0"
        type: string
    required:
    - task_type
    type: object
...
  user.updateUserStatusRequest:
    properties:
      is_active:
        type: boolean
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO Credit
  version: 1.0.0
paths:
...
  /api/v1/admin/tasks/dispatch:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.DispatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - admin
...
  /api/v1/admin/users:
    get:
      parameters:
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - in: query
        name: user_id
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - admin
  /api/v1/admin/users/{id}/status:
    put:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.updateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - admin
  /api/v1/admin/users/{id}/switch:
    post:
      parameters:
      - description: 目标账号ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - admin
...
  /api/v1/dashboard/stats/daily:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询天数，最大7天
        in: query
        name: days
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      summary: 获取每日收支统计
      tags:
      - dashboard
  /api/v1/dashboard/stats/top-customers:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询天数，最大7天
        in: query
        name: days
        required: true
        type: integer
      - description: 返回数量，最大10
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      summary: 获取Top客户
      tags:
      - dashboard
...
  /api/v1/leaderboard:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      summary: 获取排行榜列表
      tags:
      - leaderboard
...
  /api/v1/leaderboard/users/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      summary: 获取指定用户排名
      tags:
      - leaderboard
...
swagger: "2.0"